refactor(admin): migrate src/pages/ to src/features/<bc>/<resource>/
bulletproof-react layout — every page moves under a feature folder
named after the BC it belongs to:

  src/pages/products/{list,create,edit,form}.tsx
    -> src/features/catalog/products/{list,create,edit,form}.tsx
  src/pages/login.tsx
    -> src/features/identity/auth/login.tsx
  src/pages/coming-soon.tsx
    -> src/features/_shared/coming-soon.tsx

App.tsx imports re-routed through the same `@/*` path alias — no Refine
resource definitions or React Router paths change, only the file
locations. Existing Playwright E2E spec keeps passing because URL paths
are untouched.

`@/pages/*` is now empty and has been removed; ESLint-style import
zoning per feature lands in the same wave as the API-typed forms (RF-27)
once @pim/shared-types is generated from the OpenAPI spec.

Refs #176
