+++ apps/admin/src/features/catalog/products/form.tsx
