+++ apps/admin/src/features/catalog/products/list.tsx
